Add AdminMenu component and role-based nav
