+++ backend/app/main.py
@@ -1,0 +1,1 @@
+"""Placeholder module so tests can import backend.app.main."""
