+++ frontend/src/AdminPending.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import AdminMenu from './AdminMenu';
 import api from './api';
 import './AdminPending.css';
 
@@ -8,9 +8,7 @@ function AdminPending() {
   // Temporary toast message shown when a user is approved
   const [toast, setToast] = useState('');
   const [error, setError] = useState('');
-  const [menuOpen, setMenuOpen] = useState(false);
   const [selectedRoles, setSelectedRoles] = useState({});
-  const navigate = useNavigate();
 
   const token = localStorage.getItem('token');
 
@@ -71,10 +69,6 @@ function AdminPending() {
     }
   };
 
-  const handleLogout = () => {
-    localStorage.removeItem('token');
-    navigate('/login');
-  };
 
   if (!token) {
     return <p className="pending-container">Login required.</p>;
@@ -82,21 +76,7 @@ function AdminPending() {
 
   return (
     <div className="pending-container">
-      <div className="admin-menu">
-        <button
-          className="menu-button"
-          onClick={() => setMenuOpen((open) => !open)}
-        >
-          Admin Menu
-        </button>
-        {menuOpen && (
-          <div className="dropdown-menu">
-            <Link to="/dashboard">Dashboard</Link>
-            <Link to="/admin/jobs">Job Matching</Link>
-            <button onClick={handleLogout}>Logout</button>
-          </div>
-        )}
-      </div>
+      <AdminMenu />
       <h2>Pending Registrations</h2>
       {toast && <div className="toast">{toast}</div>}
       {error && <p className="error">{error}</p>}
